requirements.txt
bencoder
psycopg2
libtorrent
redis
requests
parse-torrent-name
bencodepy
aiodns
websocket-client

